Comment out shims-vue.d.ts
I had an error in my app when trying to import a component from
a .vue file that did not exist. However, typescript and my IDE did
not catch the error because this shim was saying that any .vue
file is okay even if it doesn't exist? So wanna remove it for now,
not sure the purpose of it in the first place.