Create edit flashcard page
The edit flashcard page allows the user to edit any existing flashcards.
Since there was already a create flashcard page with a form exactly
like what I needed, I moved that code into its own FlashcardForm
component and added an id prop to it. Using the id prop, the
FlashcardForm would either display an edit form with the existing
data or if the id prop was not passed in display a form to add
a new flashcard.