Create an updateAllDocs script to bulk operations
I am currently in the process of revamping how the firestore data
is fetched from firebase on the client. Specifically, I want to
drastically reduce the number of reads a user has. The strategy
I am planning on implementing is to cache all documents and when
a user revisits the site only fetch the documents that has been
modified more recently than the most recently cached document has
been modified. To accomplish this, I am going to store a
lastModified field in every document. This will allow me to fetch
all documents whose lastModified field is more recent than an
existing lastModified field.