Add the ability to create unit tests
Quasar had a command I could run to configure jest within the
project. I ran that command and had to manually install a few
packages and then unit testing was up and working. Eventually
I'll want to remove the example unit test files but for now they
are fine to keep as examples.